Move tests for add
